microservices/mcp-server/requirements.txt
fastmcp
openai
uvicorn[standard]
httpx
python-dotenv
